Migrate to network config service
