+++ packages/apps/dashboard/server/.gitignore
@@ -36,3 +36,6 @@ lerna-debug.log*
 
 # Redis Data
 /redis_data
+
+.env.development
+.env.production
